Moved Colors enum.
